Moving assignors tests out
